+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - sync task baseDir with project baseDir to handle migrated tasks
 - correcly showing the project files with special characters
 - increased timeout of semantic search tool to avoid errors when the tool is running multiple times in parallel
+- added Brainstorming as another BMAD workflow option
 
 ## [0.52.0]